[ALPS05608524] serial: 8250: using uart new map
In order to avoid seldom uart incorrect baudrate issue which caused HWT,
change to use uart new map reg.
Because in current solution, DLL is using the same address of TX/RX with
LSR in A mode. It is possible to be overwritten when unexpected TX/RX
occurs when LSR being in A mode.

MTK-Commit-Id: e7f73f8f55c8f507668ffb0e14005f73a8577a7e

Change-Id: Ia699e717afed6d1576f4c5fbe4ba62233c59ebb6
Signed-off-by: huihui.wang <huihui.wang@mediatek.com>
CR-Id: ALPS05608524
Feature: UART(uart)